Added base popup activity to manage common things between popup activities
