+++ app/src/main/java/blockeq/com/stellarwallet/activities/MyWalletActivity.kt
@@ -1,20 +1,15 @@
 package blockeq.com.stellarwallet.activities
 
-import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
 import blockeq.com.stellarwallet.R
-import kotlinx.android.synthetic.main.activity_my_wallet.*
 
-class MyWalletActivity : AppCompatActivity() {
+class MyWalletActivity : BasePopupActivity() {
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_my_wallet)
-        supportActionBar?.setDisplayShowTitleEnabled(false)
-        cancelButton.setOnClickListener {
-            finish()
-            overridePendingTransition(R.anim.stay, R.anim.slide_out_down)
-        }
+    override fun setTitle(): Int {
+        return R.string.title_activity_my_wallet
+    }
+
+    override fun setContent(): Int {
+        return R.layout.content_my_wallet
     }
 
 }
