+++ app/src/main/java/blockeq/com/stellarwallet/activities/WalletsActivity.kt
@@ -1,20 +1,15 @@
 package blockeq.com.stellarwallet.activities
 
-import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
 import blockeq.com.stellarwallet.R
-import kotlinx.android.synthetic.main.activity_wallets.*
 
-class WalletsActivity : AppCompatActivity() {
+class WalletsActivity : BasePopupActivity() {
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_wallets)
-        supportActionBar?.setDisplayShowTitleEnabled(false)
-        cancelButton.setOnClickListener {
-            finish()
-            overridePendingTransition(R.anim.stay, R.anim.slide_out_down)
-        }
+    override fun setTitle(): Int {
+        return R.string.title_activity_wallets
+    }
+
+    override fun setContent(): Int {
+        return R.layout.content_wallets
     }
 
 }
