+++ app/src/main/res/layout/activity_base_popup.xml
@@ -2,6 +2,7 @@
 <android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".activities.WalletsActivity">
@@ -16,21 +17,21 @@
             android:layout_width="match_parent"
             android:layout_height="?attr/actionBarSize"
             android:background="?attr/colorPrimary"
+            app:contentInsetStart="0dp"
             app:popupTheme="@style/AppTheme.PopupOverlay"
-            app:theme="@style/ToolbarTheme"
-            app:contentInsetStart="0dp">
+            app:theme="@style/ToolbarTheme">
 
             <RelativeLayout
                 android:layout_width="match_parent"
                 android:layout_height="match_parent">
 
                 <TextView
+                    android:id="@+id/titleText"
                     style="@style/TitleText"
                     android:layout_width="wrap_content"
                     android:layout_height="match_parent"
                     android:layout_centerInParent="true"
-                    android:gravity="center"
-                    android:text="@string/title_activity_wallets" />
+                    android:gravity="center" />
 
                 <ImageButton
                     android:id="@+id/cancelButton"
